+++ src/AgentFrameworkToolkit.AzureOpenAI/AzureOpenAIConnection.cs
@@ -1,4 +1,5 @@
 ﻿using Azure.AI.OpenAI;
+using Azure.Core;
 
 namespace AgentFrameworkToolkit.AzureOpenAI;
 
@@ -13,9 +14,14 @@ public class AzureOpenAIConnection
     public required string Endpoint { get; set; }
 
     /// <summary>
-    /// The API Key
+    /// The API Key (or use Credentials instead for RBAC)
     /// </summary>
     public string? ApiKey { get; set; }
+
+    /// <summary>
+    /// Credentials for Role-Based Access Control (Example 'DefaultAzureCredential' or 'AzureCliCredential') [or use ApiKey instead]
+    /// </summary>
+    public TokenCredential? Credentials { get; set; }
 
     /// <summary>
     /// An Action that allow you to set additional options on the AzureOpenAIClientOptions
@@ -26,6 +32,4 @@ public class AzureOpenAIConnection
     /// The timeout value of the LLM Call (if not defined the underlying infrastructure's default will be used)
     /// </summary>
     public TimeSpan? NetworkTimeout { get; set; }
-
-    //todo - Support RBAC
 }