+++ src/AgentFrameworkToolkit/AgentFrameworkToolkitException.cs
@@ -1,0 +1,8 @@
+﻿namespace AgentFrameworkToolkit;
+
+/// <summary>
+/// An AgentFrameworkToolkit Exception 
+/// </summary>
+/// <param name="message">Error Message</param>
+/// <param name="innerException">Inner Exception</param>
+public class AgentFrameworkToolkitException(string message, Exception? innerException = null) : Exception(message, innerException);